requirements.txt
pygame>=2.5
esper>=2.5
PyYAML>=6.0

